Context menu implementation complete
Currently uses long clicks to open the menu. It launches all the add
CRUD activities fine. I also added manifests for the edit activities. It
tries to launch them but it fails. (not my fault??) We can remove the
menu options below that launch them normally. May want to have something
to indicate to user they can get the context menu by long clicking.
Another problem is when creating a new list it needs an item to be able
to open the menu at this point.
#24 #31 #32 #42 #43 #45 #46