#29 #37 random item finished, fixes and cleans
Toggle buttons are defaulted to true in dummyDB() now. They also set
their proper value upon starting the app. The randomize function for
items works with and without exclusion and covers cases of 0, 1, and 2+
items. Also renamed the context menu title. Has a tiny prototype for the
list randomization which can be easily adapted to mirror the item one
once a method to tab selection and a list possible tabs is implemented.